IdentityApp
In this web application, which provides user and role management with ASP.NET Core Identity, e-mail confirmation is required for registration and users can reset their password via e-mail with the "Forgot My Password" option.
There is a users and roles page and
Only the admin user can see the roles and users section and change users' roles.